Change maximum get attempts to just a single byte
There is hardly ever a case this needs to be larger than a couple of
dozen. By using a single byte we can make lossy map entries a bit
smaller.

+++ cmd/bonanza_storage_shard/main.go
@@ -85,7 +85,7 @@ func main() {
 				},
 				configuration.LeasesMapRecordsCount,
 				leaseComparator,
-				configuration.LeasesMapMaximumGetAttempts,
+				uint8(configuration.LeasesMapMaximumGetAttempts),
 				int(configuration.LeasesMapMaximumPutAttempts),
 				"LeasesMap",
 			),
